requirements.txt
fastapi
pydantic
uvicorn
sqlalchemy
pytest
requests
passlib
argon2_cffi
bcrypt
python-jose[cryptography]
python-multipart